pyproject.toml
[tool.poetry]
name = "zflux"
version = "0.1.0"
description = ""
authors = ["Benedikt Kristinsson <benedikt@lokun.is>"]

[tool.poetry.dependencies]
python = "^3.8"
zmq = "^0.0.0"
influxdb = "^5.3.0"
loguru = "^0.5.3"
requests = "^2.24.0"
pyyaml = "^5.3.1"

[tool.poetry.scripts]
connect = "zflux.zflux:main"
stress = "zflux.stresstester:main"
wtf = "zflux.testthis:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
loguru = "^0.5.3"
zmq = "^0.0.0"
requests = "^2.24.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
